Added Enemy class, EnemyFactory, EnemyLoader
Сделал класс врага, тхтшник для них, парсер, фактори для создания врагов, проверил загрузку через тесты в мейне. Пока вроде все работает.

+++ Blade&Kami/EnemyData.h
@@ -5,7 +5,7 @@
 
 struct Spell {
     std::string name; //спеллы переделаю
-    float power;
+    double power;
     int cost;
 };
 
@@ -14,10 +14,10 @@ struct EnemyData {
     int strength;
     int agility;
     int spirit;
-    float evasion;
-    float accuracy;
+    double evasion;
+    double accuracy;
     int minDamage;
     int maxDamage;
-    float maxHealth;
+    double maxHealth;
     std::vector<Spell> spells;
 };
